+++ Assets/InfiniteScrollView/Scripts/Runtime/VerticalInfiniteScrollView.cs
@@ -1,15 +1,15 @@
 ﻿using Cysharp.Threading.Tasks;
 using UnityEngine;
 
-namespace HowTungTung
+namespace InfiniteScrollViews
 {
     public class VerticalInfiniteScrollView : InfiniteScrollView
     {
         public float spacing;
 
-        public override void RefreshCellVisibility()
+        protected override void RefreshCellVisibility()
         {
-            if (dataList.Count == 0)
+            if (_dataList.Count == 0)
                 return;
 
             // Viewport
@@ -24,25 +24,25 @@ namespace HowTungTung
 
             // Hide
             float contentHeight = padding.top;
-            for (int i = 0; i < dataList.Count; i++)
+            for (int i = 0; i < _dataList.Count; i++)
             {
-                var visibleRange = new Vector2(contentHeight, contentHeight + dataList[i].cellSize.y);
+                var visibleRange = new Vector2(contentHeight, contentHeight + _dataList[i].cellSize.y);
                 if (visibleRange.y < viewportRange.x || visibleRange.x > viewportRange.y)
                 {
                     RecycleCell(i);
                 }
-                contentHeight += dataList[i].cellSize.y + spacing;
+                contentHeight += _dataList[i].cellSize.y + spacing;
             }
 
             // Show
             contentHeight = padding.top;
-            for (int i = 0; i < dataList.Count; i++)
+            for (int i = 0; i < _dataList.Count; i++)
             {
-                var visibleRange = new Vector2(contentHeight, contentHeight + dataList[i].cellSize.y);
+                var visibleRange = new Vector2(contentHeight, contentHeight + _dataList[i].cellSize.y);
                 if (visibleRange.y >= viewportRange.x && visibleRange.x <= viewportRange.y)
                 {
                     InfiniteCell cell = null;
-                    if (cellList[i] == null)
+                    if (_cellList[i] == null)
                     {
                         if (_cellPool.Count > 0) cell = _cellPool.Dequeue();
                         else Debug.Log("<color=#ff4242>The cell display error occurred, not enough cells in the cell pool!!!</color>");
@@ -52,11 +52,11 @@ namespace HowTungTung
                     if (cell != null) dirCoeff = cell.RectTransform.pivot.y > 0 ? -1f : 1f;
                     SetupCell(cell, i, new Vector2(padding.left - padding.right, contentHeight * dirCoeff));
                     if (visibleRange.y >= viewportRange.x)
-                        cellList[i]?.transform.SetAsLastSibling();
+                        _cellList[i]?.transform.SetAsLastSibling();
                     else
-                        cellList[i]?.transform.SetAsFirstSibling();
+                        _cellList[i]?.transform.SetAsFirstSibling();
                 }
-                contentHeight += dataList[i].cellSize.y + spacing;
+                contentHeight += _dataList[i].cellSize.y + spacing;
             }
 
             // Check scroll position
@@ -74,52 +74,99 @@ namespace HowTungTung
             }
         }
 
-        public sealed override async UniTask Refresh()
+        public sealed override void Refresh(bool disabledRefreshCells = true)
         {
-            if (!IsInitialized)
-            {
-                await InitializePool();
-            }
+            if (!IsInitialized()) return;
+
             if (scrollRect.viewport.rect.height == 0)
             {
-                await DelayToRefresh();
+                this.DelayToRefresh(disabledRefreshCells).Forget();
             }
             else
             {
-                DoRefresh();
+                this.DoRefresh(disabledRefreshCells);
             }
         }
 
-        private void DoRefresh()
+        protected sealed override void DoRefresh(bool disabledRefreshCells)
         {
             if (scrollRect == null) return;
 
-            float height = padding.top;
-            for (int i = 0; i < dataList.Count; i++)
+            if (!disabledRefreshCells)
             {
-                height += dataList[i].cellSize.y + spacing;
+                // Refresh content size
+                float height = padding.top;
+                for (int i = 0; i < _dataList.Count; i++)
+                {
+                    height += _dataList[i].cellSize.y + spacing;
+                }
+                height += padding.bottom;
+                scrollRect.content.sizeDelta = new Vector2(scrollRect.content.sizeDelta.x, height);
+
+                // Recycle all cells first
+                for (int i = 0; i < _cellList.Count; i++)
+                {
+                    RecycleCell(i);
+                }
+
+                // Refresh cells view
+                this.RefreshCellVisibility();
+
+                // Invoke onRefresh callback
+                onRefreshed?.Invoke();
             }
-            for (int i = 0; i < cellList.Count; i++)
-            {
-                RecycleCell(i);
-            }
-            height += padding.bottom;
-            scrollRect.content.sizeDelta = new Vector2(scrollRect.content.sizeDelta.x, height);
-            this.RefreshCellVisibility();
-            onRefreshed?.Invoke();
+            // Mark flag for refresh at next scrolling
+            else this._disabledRefreshCells = true;
         }
 
-        private async UniTask DelayToRefresh()
+        protected sealed override async UniTask DelayToRefresh(bool disabledRefreshCells)
         {
             await UniTask.Yield(PlayerLoopTiming.LastPostLateUpdate);
-            DoRefresh();
+            DoRefresh(disabledRefreshCells);
+        }
+
+        protected override void RefreshAndCheckVisibleInfo()
+        {
+            // Reset visible count
+            this.visibleCount = 0;
+
+            // Viewport
+            float viewportInterval = scrollRect.viewport.rect.height;
+
+            // Check content direction pivot
+            if (this._contentDirCoeff == 0) this._contentDirCoeff = scrollRect.content.pivot.y > 0 ? 1f : -1f;
+
+            // Set content direction
+            float minViewport = scrollRect.content.anchoredPosition.y * this._contentDirCoeff;
+            Vector2 viewportRange = new Vector2(minViewport - extendVisibleRange, minViewport + viewportInterval + extendVisibleRange);
+
+            // Show
+            float contentHeight = padding.top;
+            for (int i = 0; i < _dataList.Count; i++)
+            {
+                var visibleRange = new Vector2(contentHeight, contentHeight + _dataList[i].cellSize.y);
+                if (visibleRange.y >= viewportRange.x && visibleRange.x <= viewportRange.y)
+                {
+                    // Calcuate visible count
+                    this.visibleCount++;
+
+                    // Check filled flag
+                    if (_cellList[i] == null) this.isVisibleRangeFilled = false;
+                    else this.isVisibleRangeFilled = true;
+                }
+                contentHeight += _dataList[i].cellSize.y + spacing;
+            }
+
+            // Adjust filled flag while cell removing
+            if (this.visibleCount < this.lastMaxVisibleCount) this.isVisibleRangeFilled = false;
+            this.lastMaxVisibleCount = this.visibleCount;
         }
 
         public override void Snap(int index, float duration)
         {
-            if (!IsInitialized)
+            if (!IsInitialized())
                 return;
-            if (index >= dataList.Count ||
+            if (index >= _dataList.Count ||
                 index < 0)
                 return;
             if (scrollRect.content.rect.height < scrollRect.viewport.rect.height)
@@ -127,10 +174,10 @@ namespace HowTungTung
             float height = padding.top;
             for (int i = 0; i < index; i++)
             {
-                height += dataList[i].cellSize.y + spacing;
+                height += _dataList[i].cellSize.y + spacing;
             }
 
-            height = this.CalculateSnapPos(ScrollType.Vertical, this.snapAlign, height, dataList[index]);
+            height = this.CalculateSnapPos(ScrollType.Vertical, this.snapAlign, height, _dataList[index]);
 
             if (scrollRect.content.anchoredPosition.y != height)
             {
@@ -139,15 +186,18 @@ namespace HowTungTung
             }
         }
 
-        public override async UniTask Remove(int index, bool withRefresh = true)
+        public override bool Remove(int index, bool withRefresh = true)
         {
-            if (index >= dataList.Count ||
+            if (!this.IsInitialized())
+                return false;
+            if (index >= _dataList.Count ||
                 index < 0)
-                return;
+                return false;
 
-            var removeCell = dataList[index];
-            await base.Remove(index, withRefresh);
+            var removeCell = _dataList[index];
+            bool result = base.Remove(index, withRefresh);
             scrollRect.content.anchoredPosition -= new Vector2(0, removeCell.cellSize.y + spacing);
+            return result;
         }
     }
 }